pyproject.toml
[project]
name = "attention-branch-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "huggingface-hub>=0.35.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "tensorboardx>=2.6.4",
    "torchvision>=0.23.0",
    "transformers>=4.57.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.1.0",
]
